Create Convert_secs_to_mins\hours
Learned a lot in this one. I learned about the %5d piece of syntax you can include in the print function (in quotes) that will tell C++ that the next piece of data should display decimal integer with up to 5 characters(or however many characters you decide to specify with the number that comes after the % sign). I learned that the fflush() function flushes out the keyboard cache or keyboard buffer which is useful when you're working with input. After it flushes the getchar() function pauses the program waiting for the user to enter a key. I think since it's empty that means any key can be pressed, which will cause the program to resume. Since there isn't really anything after that it basically just works as a mechanism that pauses the program right before it's completion so that the user has to press a button for it to close. 
I also played around with the time() function which the book said is used to help generate random numbers with the srand() function (Not to be confused with rand(), which is only kinda random, it seems like when rand() generates random number it only generates them once and then you're stuck with those numbers forever after that).
Anyway wow, that was a lot. It was cool and fun to learn about this stuff. Time to move on to the next one.